requirements.txt
python-dotenv
django
cloudinary
pillow
djangorestframework 
djangorestframework-simplejwt
gunicorn==21.2.0
psycopg2-binary
django-cors-headers==4.6.0
sqlparse==0.5.1
virtualenv==20.26.3
requests
django-role-permissions
daphne 
django-filters
watchfiles
#add more as needed